requirements.txt
agora-token-builder==1.0.0
asgiref==3.4.1
autopep8==1.6.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
cloudinary==1.29.0
colorama==0.4.5
commonmark==0.9.1
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.1
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.11.0
django-crispy-forms==1.7.2
django-environ==0.9.0
django-extensions==3.1.5
django-filter==21.1
django-heroku==0.3.1
django-on-heroku==1.1.2
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
filelock==3.7.0
gunicorn==20.1.0
httpie==3.2.1
idna==3.3
importlib-metadata==4.11.4
is-safe-url==1.0
Markdown==3.3.7
multidict==6.0.2
numpy==1.22.3
paytmchecksum==1.7.0
Pillow==9.1.0
platformdirs==2.5.2
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycryptodome==3.15.0
Pygments==2.12.0
PyJWT==2.3.0
PySocks==1.7.1
python-decouple==3.6
pytz==2021.3
requests==2.27.1
requests-toolbelt==0.9.1
rich==12.4.4
scipy==1.8.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
urllib3==1.26.9
virtualenv==20.14.1
whitenoise==6.2.0
zipp==3.8.0